Day 7
I did some refactoring of the io module system. It now uses
processes instead of callbacks. This allows a little more control
over read and write interrupts in the intcode machine. This means
that some earlier solutions had to be adjusted (technically not a
requirement but I'd prefer to keep everything working).