Day 9
Relative addressing mode added and option to explicitly
accumulate output on DiagnosticSystem. I might rename that
to debugger at some point ...

The day's problem hinted at the program taking a few seconds on
slower machines but it was quite fast, doubly so given that it's
running on persistent data structures. Hat tip to BEAM.